Implementation of smart LoadData function
each db details will be saved as a tuple object - "constructs tuples for each db, [0] is the db name, [1] is the table object, [2] is the number of columns in db"
this way it will be easy to add future tables and db. In order to add another machine db, just create a new tuple object within the above rules, and add it to tabs array.