Creation of DB now available & tabs tuples implementation was modified
- create_database was changed to manage_database
Now functions as password manager function and calls create_database
- new create_database function was introduced which actually creates the data base :smile:  Logic:
User inputs db name, if the name is legit, the function creates the db.
Than the for loop starts to send sql queries to the curser, each query is built on the fly according to the create_tabs_tuples information. -> the connection commits the queries and close the connection. Than find_database() function is being called in order to refresh the db menus. (since a new db was added)