Introducing logging!
- Log file will be saved in local %appdata%/auto_baseline_logs.
Logs are divided into 5 levels:
DEBUG

Detailed information, typically of interest only when diagnosing problems.

INFO

Confirmation that things are working as expected.

WARNING

An indication that something unexpected happened, or indicative of some problem in the near future (e.g. ‘disk space low’). The software is still working as expected.

ERROR

Due to a more serious problem, the software has not been able to perform some function.

CRITICAL

A serious error, indicating that the program itself may be unable to continue running.
- Started changing prints to loggers, still has some work to do there.
- start_logger function starts the logger with the log '--start application--'
- closeEvent function: overrides pyqt5 default closeEvent, logging '--application closeevent- in the end of the log and closes the file.
todo: if not imported, remind the user to import and ask if you sure you want to close.
- start_logger call is in the end of main MainWindow class function calls.
- config file updated with LOG dir